% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tartan-package.R
\docType{package}
\name{tartan-package}
\alias{tartan}
\alias{tartan-package}
\title{tartan: Tartan R-Package Development Example}
\description{
Provides example code for various R-package
   techniques. Currently includes example code for writing 
   functions, documenting functions, testing, and data.
}
\author{
\strong{Maintainer}: Scotty Tartan \email{scottie.dog@example.com}

}
\keyword{internal}
