% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tartans}
\alias{tartans}
\title{Tartans Dataset}
\format{An object of class \code{character} of length 1.}
\source{
Scraped from the \href{www.tartanregister.gov.uk}{Tartan Register},
14 July 2019
}
\usage{
data(tartans)
}
\description{
A data frame with information about different Tartans in Scotland.
}
\keyword{datasets}
